+++ internal/k8s_watcher/watcher.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/bitcoin-sv/arc/pkg/blocktx"
+	"github.com/bitcoin-sv/arc/internal/blocktx"
 	"github.com/bitcoin-sv/arc/pkg/metamorph"
 )
