+++ pkg/api/merkle_roots_verifier/allow_all.go
@@ -3,7 +3,7 @@ package merklerootsverifier
 import (
 	"context"
 
-	"github.com/bitcoin-sv/arc/pkg/blocktx"
+	"github.com/bitcoin-sv/arc/internal/blocktx"
 )
 
 type allowAllMerkleRootsVerifier struct{}
