--- pkg/blocktx/client_mocks.go
@@ -1,5 +1,0 @@
-package blocktx
-
-//go:generate moq -pkg mocks -out ./mocks/merkle_roots_verifier_mock.go . MerkleRootsVerifier
-
-//go:generate moq -pkg mocks -out ./mocks/blocktx_client_mock.go . Watcher
